Implemented Truck, Box and First Fit algorithm.
The algorithm currently works as intended so far (with *limited* testing).
However, remainingHeight in Truck is currently not enforced with stacks of
boxes and is currently subtracted from with each box added to the truck
bed. Ideally the highest stack of boxes needs to be checked so that it is
less than TRUCK_HEIGHT.